--upload the csv file

+++ Bhav.Application/DTOs/BhavCopyRaw.cs
@@ -8,13 +8,13 @@ namespace Bhav.Application.DTOs
 {
     public class BhavCopyRaw
     {
-        public string SecurityCode { get; set; } = string.Empty;
-        public string SecurityName { get; set; } = string.Empty;
+        public string SecurityCode { get; set; } = string.Empty;  
+        public string SecurityName { get; set; } = string.Empty; 
         public decimal? OpenPrice { get; set; }
         public decimal? HighPrice { get; set; }
         public decimal? LowPrice { get; set; }
         public decimal? ClosePrice { get; set; }
-        public int? Volume { get; set; }  
+        public long? Volume { get; set; }       
         public decimal? TradedValue { get; set; }
         public DateTime? TradeDate { get; set; }
     }
